// src/api/types/packagesSummary.ts
export interface PackagesSummary {
        mal_id: number;
        score: number;
        title: string;
        year: number;
        images: {
            webp: {
                image_url: string;

            }
        }
        rank: number;
        rating?: string;
        type?: string;
    

}